Theme: Show video for featured workshop on lander instead of thumbnail
The mockup appears to call for the video of the featured workshop to be
embeded at the top of the lander page, rather than a featured image.

Since the workshop post type doesn't have a meta field for the video
URL, but instead just has a required video block, we can search for that
block and render it individually on the lander page.