Markdown import: Use `$supported_post_types` as a string
It appears that the class was originally written to allow for multiple
post types, and so some places in the class assume the property is an
array. However, other places in the class assume it's a string, and
it's actually defined as a string.

Since we have no plans to use this for other post types, this changes
the property name to `$supported_post_type` (singular) to be more
semantically correct, and ensures that everywhere in the class treats
it as a string instead of as an array.

Fixes #72