Plugin: Disable markdown importing
All of the markdown content from the wptrainingteam repo has been imported, and the repo is not being actively maintained at this time. Some content tasks, like removing outdated lessons, are complicated using the repo, and are much simpler via simply editing the posts in the learn.wordpress.org site.

We can revisit this later and re-enable markdown import if there is a compelling reason to do so.