Changes to featured workshop and video grid components
Update and augment the supporting functions for retrieving workshop
posts and featured workshop posts so that the components that use them
can flexibly handle the main query or custom queries.

This also increases the number of workshops shown on the homepage to 6
and updates the workshop archive so that the featured workshop does not
also show in the grid of workshops below.

It also changes the order of workshops on series archives to be oldest
first, since the order of workshops in a series is determined by
publish date.