Plugin: Add the selectWoo library and register its assets
selectWoo is a fork of select2 that has improved accessibility. We can
use it to provide UI in the workshop post editor and in the front end
workshop filtering UI to allow users to select one or multiple items
from a long list (e.g. languages) while using partial match searching
to find the items they want. The WordCamp codebase also uses selectWoo
for similar purposes, such as selecting a country or currency.