Theme: Use profiles.wordpress.org to show workshop presenter bios
This makes a user's wporg profile the canonical source for their
biography blurb when they are listed as a presenter for a workshop.
Since it relies on a db table that won't exist in local dev
environments, and because some people haven't filled out their
wporg profile fields, it falls back on the description usermeta field
if the wporg profile's "About Me" field is blank.

Fixes 106