Workshops: Adjust block template to accommodate excerpt generation
This removes all of the `core/group` blocks from the template and
reassigns their custom class names to the wrapping column blocks or
inner blocks where appropriate. A little bit of additional styling
was necessary to ensure the layout remains the same for both the old
and new markup.

With no group blocks in the mix, excerpts will now be automatically
generated for workshop posts.

Fixes #46