Home Page: Update icons, tweak layout (#132)
Changes the icon for Workshops from a slideshow to a desktop monitor, to better portray that the workshops are videos. Slides are used in lesson plans, so the intention is to avoid confusion. This also adjusts the positioning of the icons within the home page parallelograms, and updates the icons for workshops and lesson plans in the WP Admin menu to match the ones used on the home page.

Fixes #126