Workshops: Add an application form and workflow (#129)
Adds a custom form for applying to present a workshop. The form fields are based on the existing form which was built with Jetpack's Contact Form block. This form, however, doesn't require the applicant to enter info like name and email, but instead requires that they log into their wp.org account first, where that data is already available.

The form is output by a wporg-learn/workshop-application-form custom block. Styles for the form front end are added to the Learn theme stylesheet. Field validation errors are shown inline with the fields they pertain to.

On the back end, a successful form submission will create a new workshop post with the "Needs Vetting" post status, which is set up via the Edit Flow plugin. Some of the application fields are used to populate the workshop post, while others are just information for the application reviewer. All of the fields from the original application are output in a metabox at the bottom of the Edit Workshop screen.

This is the third recent project I've done that called for validation of some kind of structured data, so I decided to build a tool that can flexibly handle a lot of different use cases, so I can stop reinventing the wheel every time. The result is the `\WordPressdotorg\Validator` class, which lives in wp.org's mu-plugins. Provide the validator instance with a schema for the structured data, run some data through it, and it will output detailed information about anything invalid about the data.

Refs #23