Theme Archives: Handle views with no available posts
The `get_post_type` function returns empty on an archive page whose
query has no results, even when it's specifically a post type archive or
a related taxonomy archive. So we have to use alternative means to
determine the current post type so we can output the correct labels for
things like the search form and the "no posts found" message. This also
tweaks the spacing for the "no posts found" message so it's not crammed
right up against filters/nav above it.

Fixes #139