Workshops: Add field to link to a lesson/quiz (#150)
Adds an interface for picking a lesson to link to a workshop. The lesson's post ID is stored as postmeta. Then on the front end, the Workshop Details block (which also outputs things like the duration and the language) will render a button that links to the lesson's quiz. Note that this button is part of the Workshop Details block, which also lists the info about the workshop like length and language.

Fixes #142