Normalize the card component for use in multiple contexts
Before this, lesson plans and search results each had their own concept
of a "card" for displaying each post in a set of results. This
normalizes to one concept, based on the latest design for the lesson
plan archive, and implements that concept for both the lesson plans
archive and search results. Eventually we'll use it for courses as well.