Theme: Add filtering to lesson plan archive, improve CPT-specific search (#183)
This builds on the design for the Lesson Plans archive page in #153 to turn the links in the sidebar into filters that can be applied to change the result set. Terms from multiple taxonomies can be selected simultaneously. This also specifies an action for the search forms found on workshop, lesson plan, and course archive pages so that a 404 isn't returned when searching from an inner page of the archive results.

Fixes #182