Add a locale switcher and a notice when a chosen locale isn't fully translated (#202)
Adds a locale switcher as a wp.org mu-plugin. It lives up in the Admin Bar at the top of the screen, which shows on the front end of Learn (and other wp.org sites) even if the visitor isn't signed in. The reasoning for putting the switcher here is:

* It will be in the same place on every page. If it was lower down within the site header, it would have to be in a different place on the homepage than on internal pages.
* It is reusable across all wp.org sites, should we have other places in the future where it makes sense to use it (and we have sufficient translation).

The switcher uses WP/Gutenberg's native Combobox component for selecting the locale, rather than the select2 JavaScript library (or the selectWoo variant). The hope is that this will be both more accessible and more maintainable, since the future of development for select2/selectWoo is uncertain.

This also adds a dismissible notice to the Learn theme right beneath the site header that shows when a chosen locale has less than 90 percent translation coverage. The notice explains the situation and provides links to the translation project where a visitor could add translations for their chosen locale if they want to.

Refs #187