Modify the workshop post type archive query to prioritize chosen locale
This will sort workshops in the archive so that those with a video
language that has the same *root* locale that the user has set will
be first, followed by those with video captions of the same root locale,
followed by all the rest. This is done by modifying query clauses, and
is designed to accommodate filtering by taxonomy terms as well.

Addresses #188