Import test content from live site via REST (#249)
This adds a new import script that fetches workshop and lesson plan content from the live site for local development and testing. It relies on a server-side plugin that makes raw block content publicly available via the REST API.

Currently, it will fetch the most recent 50 `lesson-plan` and `wporg_workshop` posts. That could easily be extended to include other CPTs as needed.

There is some server-side code that will only allow exporting posts that use a limited set of blocks, in order to limit the potential for disclosing data. That set of blocks can also be tweaked as needed.