Adds a bordered-image utility class
In some cases screenshots may contain a lot of the color white.

This might cause the image to blend in with the white background, and may therefore not be
as impactful for the content.

This class can be applied to a specific image block to add a slight border around the image