Removing completion rate
I'm temporarily removing the completion rate from this data (commenting it out rather than deleting it) as the completion rate metric is not actually helpful, given how quizzes and completions work in our courses.

There is a more helpful data point of "average progress" that is in the dashboard reports, however it isn't readily available for access on the frontend. I'm also not convinced it's a useful data point either way, so I'd like to get this update live and we can figure that out later on.