2024 Theme: Improve aria labels for learning pathway 'See all' links
It is best practice to have the visible text label at the start of the accessible name. This adds speech activation, improving the likelihood that a voice activation technology will activate the right link or set of links. It is better if there is also visible disambiguation so that voice command users are able to distinguish more easily which link they are activating.

Fixes #2771